Move controller TLS logic to separate file
